+++ src/index.js
@@ -17,6 +17,7 @@ export { default as Countdown } from './components/Countdown/Countdown'
 export { default as DropDown } from './components/DropDown/DropDown'
 export { default as Field } from './components/Field/Field'
 export { default as Info } from './components/Info/Info'
+export { default as RadioInput } from './components/Input/RadioInput'
 export { default as TextInput } from './components/Input/TextInput'
 export { default as Footer } from './components/Footer/Footer'
 export { default as PreFooter } from './components/PreFooter/PreFooter'
@@ -26,6 +27,7 @@ export { default as SidePanel } from './components/SidePanel/SidePanel'
 export { default as SidePanelSeparator } from './components/SidePanel/SidePanelSeparator'
 export { default as SidePanelSplit } from './components/SidePanel/SidePanelSplit'
 export { default as RadioGroup } from './components/Radio/RadioGroup'
+export { default as RadioList } from './components/Radio/RadioList'
 export { default as Table } from './components/Table/Table'
 export { default as TableCell } from './components/Table/TableCell'
 export { default as TableHeader } from './components/Table/TableHeader'
